Move cert info extend to separate service part
